# openapi/opal-osa-integration.yaml
openapi: 3.0.0
info:
  title: OPAL Decision & Diagnostics API
  version: 1.0.0
  description: |
    Complete API specification for OPAL → OSA data flow, decision-making, and troubleshooting.

    This API maintains backward compatibility with existing OPAL integrations while providing
    enhanced decision layer capabilities with evidence-based recommendations, confidence scoring,
    and impact/effort analysis.

    **Key Features:**
    - Idempotent webhook processing for reliable data flow
    - Evidence-based decision making with confidence metrics
    - Comprehensive diagnostics and troubleshooting tools
    - Force sync and orchestration capabilities
    - Backward compatible with existing OPAL workflows
  contact:
    name: OPAL Development Team
    url: https://opal-2025.vercel.app
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://opal-2025.vercel.app
    description: Production server
paths:
...
  /api/diagnostics/last-webhook:
    get:
      tags:
        - Diagnostics
      summary: Retrieve recent webhook events for troubleshooting
      description: |
        Provides comprehensive webhook event history for debugging and monitoring.

        **Use Cases:**
        - Troubleshooting failed webhook deliveries
        - Monitoring agent execution patterns
        - Validating webhook signature processing
        - Identifying duplicate or missing events
      operationId: getWebhookDiagnostics
      parameters:
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: status
          in: query
          description: Filter events by execution status
          schema:
            type: string
            enum: [success, failure, timeout, all]
            default: all
        - name: workflow_id
          in: query
          description: Filter events by specific workflow
          schema:
            type: string
        - name: agent_id
          in: query
          description: Filter events by specific agent
          schema:
            type: string
            enum: [experiment_blueprinter, audience_suggester, content_review, roadmap_generator, integration_health, personalization_idea_generator, cmp_organizer, customer_journey, geo_audit]
        - name: hours
          in: query
          description: Number of hours to look back
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
      responses:
        '200':
          description: List of recent webhook events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookDiagnosticsResponse'
              example:
                success: true
                data:
                  events:
                    - event_id: "evt-1699876543210-abc123"
                      workflow_id: "wf-content-audit-2025"
                      agent_id: "content_review"
                      execution_status: "success"
                      received_at: "2025-11-12T15:30:00.000Z"
                      http_status: 202
                      signature_valid: true
                      processing_time_ms: 45
                      duplicate_detected: false
                    - event_id: "evt-1699876543205-def456"
                      workflow_id: "wf-audience-analysis-2025"
                      agent_id: "audience_suggester"
                      execution_status: "failure"
                      received_at: "2025-11-12T15:25:00.000Z"
                      http_status: 202
                      signature_valid: true
                      processing_time_ms: 89
                      error_details:
                        error_code: "DATA_ACCESS_ERROR"
                        error_message: "Unable to access customer data source"
                  metadata:
                    total_events: 2
                    success_rate: "50.0%"
                    time_range: "2025-11-11T15:30:00.000Z to 2025-11-12T15:30:00.000Z"
                correlation_id: "api-1699876543210-ghi789"
                timestamp: "2025-11-12T15:45:00.000Z"
                duration_ms: 156
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    OpalAgentType:
      type: string
      enum:
        - experiment_blueprinter
        - audience_suggester
        - content_review
        - roadmap_generator
        - integration_health
        - personalization_idea_generator
        - cmp_organizer
        - customer_journey
        - geo_audit
      description: Supported OPAL agent types
    ExecutionStatus:
      type: string
      enum: [success, failure, timeout]
      description: Webhook event execution status
...
    WebhookDiagnosticsResponse:
      allOf:
        - $ref: '#/components/schemas/StandardApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/WebhookEventSummary'
                metadata:
                  type: object
                  properties:
                    total_events:
                      type: integer
                      description: Total number of events in response
                    success_rate:
                      type: string
                      description: Percentage of successful events
                      example: "85.2%"
                    time_range:
                      type: string
                      description: Time range covered by results
                      example: "2025-11-11T15:30:00.000Z to 2025-11-12T15:30:00.000Z"
    WebhookEventSummary:
      type: object
      properties:
        event_id:
          type: string
          description: Unique event identifier
        workflow_id:
          type: string
          description: Associated workflow identifier
        agent_id:
          $ref: '#/components/schemas/OpalAgentType'
        execution_status:
          $ref: '#/components/schemas/ExecutionStatus'
        received_at:
          type: string
          format: date-time
          description: When the webhook was received
        http_status:
          type: integer
          description: HTTP status code returned to sender
        signature_valid:
          type: boolean
          description: Whether webhook signature was valid
        processing_time_ms:
          type: integer
          description: Time taken to process the webhook
        duplicate_detected:
          type: boolean
          description: Whether this was identified as a duplicate
        error_details:
          type: object
          description: Error information (if applicable)
          properties:
            error_code:
              type: string
            error_message:
              type: string
...
    StandardApiResponse:
      type: object
      required:
        - success
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        data:
          description: Response data (varies by endpoint)
        message:
          type: string
          description: Optional human-readable message
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Response generation timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time in milliseconds
          example: 245
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error description
          example: "Invalid request body"
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time before error occurred
          example: 25
        data:
          description: Additional error details (only in development)
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: X-Signature-SHA256
      description: |
        HMAC-SHA256 signature for webhook authentication.

        Format: sha256=<hex_digest>

        The signature is computed using the raw request body and a shared secret key.
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "VALIDATION_ERROR"
            message: "Invalid request body"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 25
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 150
...